Renamed package and made publishable
