--- packages/libs/ui/react-token-gate/README.md
@@ -1,7 +1,0 @@
-# libs-ui-react-token-gate
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test libs-ui-react-token-gate` to execute the unit tests via [Jest](https://jestjs.io).
