--- packages/libs/ui/react-token-gate/project.json
@@ -1,25 +1,0 @@
-{
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/libs/ui/react-token-gate/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "packages/libs/ui/react-token-gate/**/*.{ts,tsx,js,jsx}"
-        ]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/packages/libs/ui/react-token-gate"],
-      "options": {
-        "jestConfig": "packages/libs/ui/react-token-gate/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  }
-}
