+++ packages/libs/ui/token-gate/jest.config.ts
@@ -1,10 +1,10 @@
 /* eslint-disable */
 export default {
-  displayName: 'libs-ui-react-token-gate',
+  displayName: 'libs-ui-token-gate',
   preset: '../../../../jest.preset.js',
   transform: {
     '^.+\\.[tj]sx?$': 'babel-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../../coverage/packages/libs/ui/react-token-gate',
+  coverageDirectory: '../../../../coverage/packages/libs/ui/token-gate',
 };
