+++ packages/libs/ui/token-gate/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@quicknode/token-gate",
+  "version": "0.0.1"
+}
