+++ tsconfig.base.json
@@ -16,15 +16,13 @@
     "baseUrl": ".",
     "paths": {
       "@qn-oss/libs/api/sdk": ["packages/libs/api/sdk/src/index.ts"],
-      "@qn-oss/libs/ui/react-token-gate": [
-        "packages/libs/ui/react-token-gate/src/index.ts"
-      ],
       "@qn-oss/libs/ui/util": ["packages/libs/ui/util/src/index.ts"],
       "@qn-oss/libs/util": ["packages/libs/util/src/index.ts"],
       "@quicknode/icy-nft-hooks": [
         "packages/libs/ui/nft-react-hooks/src/index.ts"
       ],
-      "@quicknode/sdk": ["packages/libs/api/sdk/src/index.ts"]
+      "@quicknode/sdk": ["packages/libs/api/sdk/src/index.ts"],
+      "@quicknode/token-gate": ["packages/libs/ui/token-gate/src/index.ts"]
     },
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
