+++ workspace.json
@@ -8,7 +8,7 @@
     "apps-examples-token-gate-example": "packages/apps/examples/token-gate-example",
     "libs-api-sdk": "packages/libs/api/sdk",
     "libs-ui-nft-react-hooks": "packages/libs/ui/nft-react-hooks",
-    "libs-ui-react-token-gate": "packages/libs/ui/react-token-gate",
+    "libs-ui-token-gate": "packages/libs/ui/token-gate",
     "libs-ui-util": "packages/libs/ui/util",
     "libs-util": "packages/libs/util"
   }
